+++ semgrep/tests/conftest.py
@@ -1,19 +1,23 @@
 import json
 import os
+import re
 import shlex
 import tempfile
+from dataclasses import dataclass
+from functools import partial
+from io import StringIO
 from pathlib import Path
 from shutil import copytree
 from typing import Callable
 from typing import Dict
+from typing import Iterable
 from typing import List
-from typing import Mapping
 from typing import Optional
 from typing import Sequence
 from typing import Set
-from typing import Tuple
 from typing import Union
 
+import colorama
 import pytest
 from click.testing import CliRunner
 
@@ -65,7 +69,10 @@ def _clean_stdout(out):
 
 def _clean_output_json(output_json: str) -> str:
     """Make semgrep's output deterministic and nicer to read."""
-    output = json.loads(output_json)
+    try:
+        output = json.loads(output_json)
+    except json.decoder.JSONDecodeError:
+        return output_json
 
     masked_keys = [
         "tool.driver.semanticVersion",
@@ -104,49 +111,158 @@ def _clean_output_json(output_json: str) -> str:
             for skip in paths["skipped"]
         ]
 
+    # Necessary because some tests produce temp files
+    if output.get("errors"):
+        for error in output.get("errors"):
+            if error.get("spans"):
+                for span in error.get("spans"):
+                    if span.get("file"):
+                        file = span.get("file")
+                        span["file"] = file if "tmp" not in file else "tmp/masked/path"
+
     return json.dumps(output, indent=2, sort_keys=True)
 
 
-def _clean_output_sarif(output):
+def _clean_output_sarif(output_json: str) -> str:
+    try:
+        output = json.loads(output_json)
+    except json.decoder.JSONDecodeError:
+        return output_json
+
     # Rules are logically a set so the JSON list's order doesn't matter
     # we make the order deterministic here so that snapshots match across runs
     # the proper solution will be https://github.com/joseph-roitman/pytest-snapshot/issues/14
-    output["runs"][0]["tool"]["driver"]["rules"] = sorted(
-        output["runs"][0]["tool"]["driver"]["rules"],
-        key=lambda rule: str(rule["id"]),
-    )
+    try:
+        output["runs"][0]["tool"]["driver"]["rules"] = sorted(
+            output["runs"][0]["tool"]["driver"]["rules"],
+            key=lambda rule: str(rule["id"]),
+        )
+    except (KeyError, IndexError):
+        pass
 
     # Semgrep version is included in sarif output. Verify this independently so
     # snapshot does not need to be updated on version bump
-    assert output["runs"][0]["tool"]["driver"]["semanticVersion"] == __VERSION__
-    output["runs"][0]["tool"]["driver"]["semanticVersion"] = "placeholder"
+    try:
+        assert output["runs"][0]["tool"]["driver"]["semanticVersion"] == __VERSION__
+        output["runs"][0]["tool"]["driver"]["semanticVersion"] = "placeholder"
+    except (KeyError, IndexError):
+        pass
 
-    return output
+    return json.dumps(output, indent=2, sort_keys=True)
 
 
-CLEANERS: Mapping[str, Callable[[str], str]] = {
-    "--sarif": lambda s: json.dumps(_clean_output_sarif(json.loads(s))),
-    "--gitlab-sast": _clean_output_json,
-    "--gitlab-secrets": _clean_output_json,
-    "--json": _clean_output_json,
-}
+Maskers = Iterable[Union[str, re.Pattern, Callable[[str], str]]]
+
+
+def mask_capture_group(match: re.Match) -> str:
+    if not match.groups():
+        return "<MASKED>"
+    text: str = match.group()
+    for group in match.groups():
+        text = text.replace(group, "<MASKED>")
+    return text
+
+
+ALWAYS_MASK: Maskers = (
+    _clean_output_json,
+    _clean_output_sarif,
+    __VERSION__,
+    re.compile(r"python (\d+[.]\d+[.]\d+)"),
+    re.compile(r'SEMGREP_SETTINGS_FILE="(.+?)"'),
+    re.compile(r'SEMGREP_VERSION_CACHE_PATH="(.+?)"'),
+)
+
+
+@dataclass
+class SemgrepResult:
+    command: str
+    raw_stdout: str
+    raw_stderr: str
+    exit_code: int
+
+    def strip_color(self, text: str) -> str:
+        stream = StringIO()
+        desaturator = colorama.AnsiToWin32(stream, strip=True)
+        desaturator.write(text)
+        stream.seek(0)
+        return stream.read()
+
+    def mask_text(self, text: str, mask: Optional[Maskers] = None) -> str:
+        if mask is None:
+            mask = []
+        for pattern in [*mask, *ALWAYS_MASK]:
+            if isinstance(pattern, str):
+                text = text.replace(pattern, "<MASKED>")
+            elif isinstance(pattern, re.Pattern):
+                text = pattern.sub(mask_capture_group, text)
+            elif callable(pattern):
+                text = pattern(text)
+        return text
+
+    @property
+    def stdout(self) -> str:
+        return self.mask_text(self.raw_stdout)
+
+    @property
+    def stderr(self) -> str:
+        return self.mask_text(self.raw_stderr)
+
+    def as_snapshot(self, mask: Optional[Maskers] = None):
+        stdout = self.mask_text(self.raw_stdout, mask)
+        stderr = self.mask_text(self.raw_stderr, mask)
+        sections = {
+            "command": self.mask_text(self.command, mask),
+            "exit code": self.exit_code,
+            "stdout - plain": self.strip_color(stdout),
+            "stderr - plain": self.strip_color(stderr),
+            "stdout - color": stdout,
+            "stderr - color": stderr,
+        }
+        if (
+            sections["stdout - plain"] == sections["stdout - color"]
+            and sections["stderr - plain"] == sections["stderr - color"]
+        ):
+            del sections["stdout - color"]
+            del sections["stderr - color"]
+        return "\n\n".join(
+            f"=== {title}\n{text}\n=== end of {title}"
+            for title, text in sections.items()
+        )
+
+    def print_debug_info(self) -> None:
+        print(
+            "=== to reproduce (run with `pytest --pdb` to suspend while temp dirs exist)"
+        )
+        print(f"$ cd {os.getcwd()}")
+        print(f"$ {self.command}")
+        print("=== exit code")
+        print(self.exit_code)
+        print("=== stdout")
+        print(self.stdout)
+        print("=== stderr")
+        print(self.stderr)
+
+    def __iter__(self):
+        """For backwards compat with usages like `stdout, stderr = run_semgrep(...)`"""
+        yield self.stdout
+        yield self.stderr
 
 
 def _run_semgrep(
     config: Optional[Union[str, Path, List[str]]] = None,
     *,
-    target_name: str = "basic",
+    target_name: Optional[str] = "basic",
     options: Optional[List[Union[str, Path]]] = None,
-    output_format: OutputFormat = OutputFormat.JSON,
+    output_format: Optional[OutputFormat] = OutputFormat.JSON,
     strict: bool = True,
     quiet: bool = False,
     env: Optional[Dict[str, str]] = None,
     assert_exit_code: Union[None, int, Set[int]] = 0,
-    settings_file: Optional[str] = None,
     force_color: Optional[bool] = None,
     assume_targets_dir: bool = True,  # See e2e/test_dependency_aware_rule.py for why this is here
     force_metrics_off: bool = True,
-) -> Tuple[str, str]:
+    stdin: Optional[str] = None,
+) -> SemgrepResult:
     """Run the semgrep CLI.
 
     :param config: what to pass as --config's value
@@ -157,10 +273,6 @@ def _run_semgrep(
     :param settings_file: what setting file for semgrep to use. If None, a random temp file is generated
                           with default params ("has_shown_metrics_notification: true")
     """
-
-    # If delete_setting_file is false and a settings file doesnt exist, put a default
-    # as we are not testing said setting. Note that if Settings file exists we want to keep it
-    # Use a unique settings file so multithreaded pytest works well
     env = {} if not env else env.copy()
 
     if force_color:
@@ -169,13 +281,20 @@ def _run_semgrep(
     if "SEMGREP_USER_AGENT_APPEND" not in env:
         env["SEMGREP_USER_AGENT_APPEND"] = "pytest"
 
-    if not settings_file:
+    # If delete_setting_file is false and a settings file doesnt exist, put a default
+    # as we are not testing said setting. Note that if Settings file exists we want to keep it
+    # Use a unique settings file so multithreaded pytest works well
+    if "SEMGREP_SETTINGS_FILE" not in env:
         unique_settings_file = tempfile.NamedTemporaryFile().name
         Path(unique_settings_file).write_text("has_shown_metrics_notification: true")
 
         env["SEMGREP_SETTINGS_FILE"] = unique_settings_file
-    else:
-        env["SEMGREP_SETTINGS_FILE"] = settings_file
+    if "SEMGREP_VERSION_CACHE_PATH" not in env:
+        env["SEMGREP_VERSION_CACHE_PATH"] = tempfile.TemporaryDirectory().name
+    if "SEMGREP_ENABLE_VERSION_CHECK" not in env:
+        env["SEMGREP_ENABLE_VERSION_CHECK"] = "0"
+    if force_metrics_off and "SEMGREP_SEND_METRICS" not in env:
+        env["SEMGREP_SEND_METRICS"] = "off"
 
     if options is None:
         options = []
@@ -185,10 +304,6 @@ def _run_semgrep(
 
     if quiet:
         options.append("--quiet")
-
-    options.append("--disable-version-check")
-    if force_metrics_off:
-        options.append("--metrics=off")
 
     if config is not None:
         if isinstance(config, list):
@@ -208,35 +323,35 @@ def _run_semgrep(
     elif output_format == OutputFormat.SARIF:
         options.append("--sarif")
 
-    target = Path("targets") / target_name if assume_targets_dir else Path(target_name)
-    args = " ".join(shlex.quote(str(c)) for c in [*options, target])
+    targets = []
+    if target_name is not None:
+        targets.append(
+            Path("targets") / target_name if assume_targets_dir else Path(target_name)
+        )
+    args = " ".join(shlex.quote(str(c)) for c in [*options, *targets])
+    env_string = " ".join(f'{k}="{v}"' for k, v in env.items())
 
     runner = CliRunner(env=env, mix_stderr=False)
-    result = runner.invoke(cli, args)
-
-    # some helpful output for reproducing issues
-    env_string = " ".join(f'{k}="{v}"' for k, v in env.items())
-    print(f"### COMMANDS")
-    print(f"$ cd {os.getcwd()}")
-    print(f"$ {env_string} semgrep {args}")
-    print("### STDOUT")
-    print(result.stdout)
-    print("### STDERR")
-    print(result.stderr)
-    print("### END\n")
+    click_result = runner.invoke(cli, args, input=stdin)
+    result = SemgrepResult(
+        f"{env_string} semgrep {args}",
+        click_result.stdout,
+        click_result.stderr,
+        click_result.exit_code,
+    )
+    result.print_debug_info()
 
     if isinstance(assert_exit_code, set):
         assert result.exit_code in assert_exit_code
     elif isinstance(assert_exit_code, int):
         assert result.exit_code == assert_exit_code
 
-    stdout = (
-        _clean_output_json(result.stdout)
-        if result.stdout and output_format.is_json()
-        else result.stdout
-    )
+    return result
 
-    return stdout, result.stderr
+
+@pytest.fixture
+def run_semgrep():
+    yield partial(_run_semgrep, strict=False, target_name=None, output_format=None)
 
 
 @pytest.fixture