+++ semgrep/tests/e2e/test_metavariable_regex_const_prop.py
@@ -7,6 +7,6 @@ def test_metavariable_regex_const_prop(run_semgrep_in_tmp, snapshot):
         run_semgrep_in_tmp(
             "rules/metavariable-regex-const-prop.yaml",
             target_name="metavariable_propagation/metavariable-regex-const-prop.dockerfile",
-        )[0],
+        ).stdout,
         "results.json",
     )