+++ semgrep/tests/e2e/test_severity.py
@@ -3,9 +3,9 @@ import pytest
 
 @pytest.mark.kinda_slow
 def test_severity_error(run_semgrep_in_tmp, snapshot):
-    json_str = run_semgrep_in_tmp("rules/inside.yaml", options=["--severity", "ERROR"])[
-        0
-    ]
+    json_str = run_semgrep_in_tmp(
+        "rules/inside.yaml", options=["--severity", "ERROR"]
+    ).stdout
     assert json_str != ""
     assert '"severity": "INFO"' not in json_str
     assert '"severity": "WARNING"' not in json_str
@@ -15,7 +15,7 @@ def test_severity_error(run_semgrep_in_tmp, snapshot):
 def test_severity_info(run_semgrep_in_tmp, snapshot):
     # Shouldn't return errors or results, since inside.yaml has 'severity: ERROR'
     snapshot.assert_match(
-        run_semgrep_in_tmp("rules/inside.yaml", options=["--severity", "INFO"])[0],
+        run_semgrep_in_tmp("rules/inside.yaml", options=["--severity", "INFO"]).stdout,
         "results.json",
     )
 
@@ -24,7 +24,9 @@ def test_severity_info(run_semgrep_in_tmp, snapshot):
 def test_severity_warning(run_semgrep_in_tmp, snapshot):
     # Shouldn't return errors or results, since inside.yaml has 'severity: ERROR'
     snapshot.assert_match(
-        run_semgrep_in_tmp("rules/inside.yaml", options=["--severity", "WARNING"])[0],
+        run_semgrep_in_tmp(
+            "rules/inside.yaml", options=["--severity", "WARNING"]
+        ).stdout,
         "results.json",
     )
 
@@ -36,6 +38,6 @@ def test_severity_multiple(run_semgrep_in_tmp, snapshot):
     snapshot.assert_match(
         run_semgrep_in_tmp(
             "rules/inside.yaml", options=["--severity", "INFO", "--severity", "WARNING"]
-        )[0],
+        ).stdout,
         "results.json",
     )